+++ src/brushes/circle_brush.class.ts
@@ -1,5 +1,4 @@
-import { fabric } from '../../HEADER';
-import { Color } from '../color';
+import { Color } from '../color/color.class';
 import { Point } from '../point.class';
 import { Shadow } from '../shadow.class';
 import { Circle } from '../shapes/circle.class';
@@ -144,5 +143,3 @@ export class CircleBrush extends BaseBrush {
     return pointerPoint;
   }
 }
-
-fabric.CircleBrush = CircleBrush;