+++ src/canvas/canvas.class.ts
@@ -1,5 +1,6 @@
-import { fabric } from '../../HEADER';
-import { dragHandler, getActionFromCorner } from '../controls/actions';
+import { getEnv } from '../env';
+import { dragHandler } from '../controls/drag';
+import { getActionFromCorner } from '../controls/util';
 import { Point } from '../point.class';
 import { FabricObject } from '../shapes/Object/FabricObject';
 import {
@@ -515,7 +516,7 @@ export class SelectableCanvas<
     this._createUpperCanvas();
     // @ts-ignore
     this._initEventListeners();
-    this._initRetinaScaling();
+    this._isRetinaScaling() && this._initRetinaScaling();
     this.calcOffset();
     this._createCacheCanvas();
   }
@@ -833,11 +834,14 @@ export class SelectableCanvas<
     if (!target) {
       return;
     }
-    let pointer = this.getPointer(e);
-    if (target.group) {
-      // transform pointer to target's containing coordinate plane
-      pointer = sendPointToPlane(pointer, target.group.calcTransformMatrix());
-    }
+    const pointer = target.group
+      ? // transform pointer to target's containing coordinate plane
+        sendPointToPlane(
+          this.getPointer(e),
+          undefined,
+          target.group.calcTransformMatrix()
+        )
+      : this.getPointer(e);
     const corner = target.__corner || '',
       control = target.controls[corner],
       actionHandler =
@@ -1272,7 +1276,7 @@ export class SelectableCanvas<
     if (this.wrapperEl) {
       return;
     }
-    const container = fabric.document.createElement('div');
+    const container = getEnv().document.createElement('div');
     container.classList.add(this.containerClass);
     this.wrapperEl = wrapElement(this.lowerCanvasEl, container);
     this.wrapperEl.setAttribute('data-fabric', 'wrapper');