+++ src/parser/elements_parser.ts
@@ -1,6 +1,5 @@
 //@ts-nocheck
-import { fabric } from '../../HEADER';
-import { Gradient } from '../gradient';
+import { Gradient } from '../gradient/gradient.class';
 import { Group } from '../shapes/group.class';
 import { Image } from '../shapes/image.class';
 import { classRegistry } from '../util/class_registry';
@@ -9,6 +8,7 @@ import {
   multiplyTransformMatrices,
   qrDecompose,
 } from '../util/misc/matrix';
+import { storage } from './constants';
 
 const ElementsParser = function (
   elements,
@@ -77,7 +77,7 @@ const ElementsParser = function (
     };
   };
 
-  proto.extractPropertyDefinition = function (obj, property, storage) {
+  proto.extractPropertyDefinition = function (obj, property, storageType) {
     const value = obj[property],
       regex = this.regexUrl;
     if (!regex.test(value)) {
@@ -86,7 +86,8 @@ const ElementsParser = function (
     regex.lastIndex = 0;
     const id = regex.exec(value)[1];
     regex.lastIndex = 0;
-    return fabric[storage][this.svgUid][id];
+    // @todo fix this
+    return storage[storageType][this.svgUid][id];
   };
 
   proto.resolveGradient = function (obj, el, property) {